+++ main/playlist_manager/playlist_manager_panel.js
@@ -1,5 +1,5 @@
 ﻿'use strict';
-//29/05/23
+//03/07/23
 
 include('..\\..\\helpers\\helpers_xxx.js');
 include('..\\..\\helpers\\helpers_xxx_properties.js');
@@ -18,7 +18,15 @@ function _panel(customBackground = false, bSetup = false) {
 		bToolbar			: ['Use toolbar mode?', true, {func: isBoolean}],
 		bButtonsBackground	: ['Use buttons background?', false, {func: isBoolean}],
 		buttonsToolbarColor	: ['Buttons\' toolbar colour', RGB(0,0,0), {func: isInt}],
-		buttonsToolbarTransparency	: ['Buttons\' toolbar transparency', 5, {func: isInt, range: [[0,100]]}]
+		buttonsToolbarTransparency	: ['Buttons\' toolbar transparency', 5, {func: isInt, range: [[0,100]]}],
+		imageBackground			: ['Image background config', JSON.stringify({
+			enabled: true, 
+			mode: 1, 
+			art: {path: '', image: null}, 
+			transparency: 60, 
+			bProportions: true, 
+			bFill: true
+		}), {func: isJSON}],
 	};
 	for (let key in panelProperties) {panelProperties[key][3] = panelProperties[key][1];}
 	setProperties(panelProperties, 'panel_');
@@ -89,11 +97,57 @@ function _panel(customBackground = false, bSetup = false) {
 		}
 		return col;
 	};
-	this.paint = (gr) => {
+	
+	this.updateImageBg = () => {
+		if (!this.imageBackground.enabled) {this.imageBackground.art.path = null; this.imageBackground.art.image = null;}
+		let handle;
+		if (this.imageBackground.mode === 0) { // Selection
+			handle = fb.GetFocusItem(true);
+		} else if (this.imageBackground.mode === 1) { // Now Playing
+			handle = fb.GetNowPlaying() || fb.GetFocusItem(true);
+		}
+		const promise = this.imageBackground.mode === 2 && this.imageBackground.art.path.length 
+			? gdi.LoadImageAsyncV2('', this.imageBackground.art.path)
+			: handle 
+				? utils.GetAlbumArtAsyncV2(void(0), handle, 0, true, false, false)
+				: Promise().reject('No handle/art');
+		promise.then((result) => {
+			if (this.imageBackground.mode === 2) {
+				this.imageBackground.art.image = result;
+			} else {
+				if (!result.image) {throw 'Image not available';}
+				this.imageBackground.art.image = result.image;
+				this.imageBackground.art.path = result.path;
+			}
+			return window.Repaint();
+		}).catch(() => {
+			this.imageBackground.art.path = null; this.imageBackground.art.image = null;
+			return window.Repaint();
+		});
+	}
+	
+	this.paintImage = (gr, limits = {x, y, w, h, offsetH}) => {
+		if (this.imageBackground.enabled && this.imageBackground.art.image) {
+			if (this.imageBackground.bFill) {
+				const prop = limits.w / limits.h - limits.offsetH;
+				const offsetX = prop > 1 ? prop * this.imageBackground.art.image.Width : 0;
+				const offsetY = prop < 1 ? prop * this.imageBackground.art.image.Height : 0;
+				gr.DrawImage(this.imageBackground.art.image, limits.x , limits.y, limits.w, limits.h, offsetX / 2, offsetY / 2, this.imageBackground.art.image.Width - offsetX / 2, this.imageBackground.art.image.Height - offsetY, 0, this.imageBackground.transparency);
+			} else {
+				let w, h;
+				if (this.imageBackground.bProportions) {w = h = Math.min(limits.w, limits.h - limits.offsetH);}
+				else {[w , h] = [limits.w, limits.h];}
+				gr.DrawImage(this.imageBackground.art.image, (window.Width - w) / 2, Math.max((limits.h - limits.y - h) / 2 + limits.y, limits.y), w, h, 0, 0, this.imageBackground.art.image.Width, this.imageBackground.art.image.Height, 0, this.imageBackground.transparency);
+			}
+		}
+	};
+	
+	this.paint = (gr, bImage = false) => {
 		const col = this.getColorBackground();
 		if (typeof col !== 'undefined') {
 			gr.FillSolidRect(0, 0, this.w, this.h, col);
 		}
+		if (bImage) {this.paintImage(gr, {y: 0, w: this.w, h: this.h, offsetH: _scale(1)});}
 	};
 
 	this.setDefault = ({all = false, buttonText = false, buttonBar = false, oldColor = null} = {}) => {
@@ -135,10 +189,12 @@ function _panel(customBackground = false, bSetup = false) {
 	this.colors.bButtonsBackground = this.properties.bButtonsBackground[1];
 	this.colors.buttonsToolbarColor = this.properties.buttonsToolbarColor[1];
 	this.colors.buttonsToolbarTransparency = this.properties.buttonsToolbarTransparency[1];
+	this.imageBackground = JSON.parse(this.properties.imageBackground[1], (key, value) => key === 'image' ? null : value);
 	this.listObjects = [];
 	this.textObjects = [];
 	this.fontChanged();
 	this.colorsChanged();
+	this.updateImageBg();
 	if (bSetup) {
 		const defaultCol = invert(this.getColorBackground());
 		this.properties.buttonsTextColor[1] = this.colors.buttonsTextColor = this.colors.bButtonsBackground ? this.colors.default.buttonsTextColor : defaultCol;