+++ playlist_manager.js
@@ -609,9 +609,29 @@ if (!list.properties.bSetup[1]) {
 
 	addEventListener('on_playback_new_track', () => { // To show playing now playlist indicator...
 		window.Repaint();
+		panel.updateImageBg();
+	});
+	
+	addEventListener('on_selection_changed', () => {
+		panel.updateImageBg();
+	});
+
+	addEventListener('on_item_focus_change', () => {
+		panel.updateImageBg();
+	});
+
+	addEventListener('on_playlist_switch', () => {
+		panel.updateImageBg();
+	});
+
+	addEventListener('on_playback_stop', (reason) => {
+		if (reason !== 2) { // Invoked by user or Starting another track
+			panel.updateImageBg();
+		}
 	});
 
 	addEventListener('on_playlists_changed', () => { // To show/hide loaded playlist indicators...
+		panel.updateImageBg();
 		window.Repaint();
 	});
 
@@ -807,6 +827,7 @@ if (!list.properties.bSetup[1]) {
 
 	addEventListener('on_playlist_items_removed', (playlistIndex, newCount, oldName = null) => {
 		const name = plman.GetPlaylistName(playlistIndex);
+		panel.updateImageBg();
 		if (!list.isAutosave(name)) {return;}
 		// Disable auto-saving on panel cache reload and ensure future update matches the right playlist
 		if (pop.isEnabled() && debouncedUpdate && playlistIndex !== -1) {setTimeout(on_playlist_items_removed, autoSaveTimer, playlistIndex, name); return;}
@@ -817,6 +838,7 @@ if (!list.properties.bSetup[1]) {
 
 	addEventListener('on_playlist_items_added', (playlistIndex, oldName = null) => {
 		const name = plman.GetPlaylistName(playlistIndex);
+		panel.updateImageBg();
 		if (!list.isAutosave(name)) {return;}
 		// Disable auto-saving on panel cache reload and ensure future update matches the right playlist
 		if (pop.isEnabled() && debouncedUpdate && playlistIndex !== -1) {setTimeout(on_playlist_items_added, autoSaveTimer, playlistIndex, name); return;}